# results/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Smart AI",
    "version": "0.1.0"
  },
  "paths": {
    "/datasets": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "List Datasets",
        "description": "List all datasets with pagination info.",
        "operationId": "list_datasets_datasets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response List Datasets Datasets Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "datasets"
        ],
        "summary": "Create Dataset",
        "description": "Create a new dataset and start pulling its data in the background.",
        "operationId": "create_dataset_datasets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDatasetSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Dataset Datasets Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/datasets/filter/{task_type}": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Filter Datasets By Task Type",
        "description": "Filter datasets by task type.",
        "operationId": "filter_datasets_by_task_type_datasets_filter__task_type__get",
        "parameters": [
          {
            "name": "task_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Filter Datasets By Task Type Datasets Filter  Task Type  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "datasets"
        ],
        "summary": "Delete Datasets By Task Type",
        "description": "Delete all datasets of a given task type.",
        "operationId": "delete_datasets_by_task_type_datasets_filter__task_type__delete",
        "parameters": [
          {
            "name": "task_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/datasets/{name}": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Get Dataset",
        "description": "Get a dataset by name, including its worker status.",
        "operationId": "get_dataset_datasets__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Dataset Datasets  Name  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "datasets"
        ],
        "summary": "Delete Dataset",
        "description": "Delete a dataset by name (use 'ALL-' to delete every dataset).",
        "operationId": "delete_dataset_datasets__name__delete",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Dataset Datasets  Name  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/datasets/{name}/status": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Get Dataset Status",
        "description": "Get a dataset's processing status.",
        "operationId": "get_dataset_status_datasets__name__status_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Dataset Status Datasets  Name  Status Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/datasets/{name}/sample": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Get Dataset Sample",
        "description": "Get a random sample of rows from a dataset.",
        "operationId": "get_dataset_sample_datasets__name__sample_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/datasets/{name}/describe": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Get Dataset Describe",
        "description": "Get descriptive statistics for a dataset.",
        "operationId": "get_dataset_describe_datasets__name__describe_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/datasets/{name}/pca": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Get Dataset Pca",
        "description": "Get a PCA projection of a dataset.",
        "operationId": "get_dataset_pca_datasets__name__pca_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/datasets/{name}/logs": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Get Dataset Logs",
        "description": "Get a dataset's logs.",
        "operationId": "get_dataset_logs_datasets__name__logs_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateDatasetSchema": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "task_type": {
            "type": "string",
            "title": "Task Type"
          },
          "features": {
            "items": {},
            "type": "array",
            "title": "Features"
          },
          "tags": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              }
            ],
            "title": "Target"
          },
          "start_date": {
            "type": "string",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "title": "End Date"
          },
          "time_start": {
            "type": "string",
            "title": "Time Start",
            "default": "00:00:00"
          },
          "time_end": {
            "type": "string",
            "title": "Time End",
            "default": "23:59:00"
          },
          "interval": {
            "type": "integer",
            "title": "Interval",
            "default": 0
          },
          "preprocessing": {
            "$ref": "#/components/schemas/PreprocessingSchema"
          },
          "n_models": {
            "type": "integer",
            "title": "N Models",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "task_type",
          "features",
          "target",
          "start_date",
          "end_date"
        ],
        "title": "CreateDatasetSchema"
      },
      "DatasetHandling": {
        "type": "string",
        "enum": [
          "NONE",
          "REMOVE",
          "MEAN",
          "MEDIAN",
          "MAX",
          "MIN",
          "NEIGHBOR_VALUE"
        ],
        "title": "DatasetHandling"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PreprocessingSchema": {
        "properties": {
          "missing_handling": {
            "$ref": "#/components/schemas/DatasetHandling",
            "default": "NEIGHBOR_VALUE"
          },
          "outlier_handling": {
            "type": "boolean",
            "title": "Outlier Handling",
            "default": false
          },
          "scale": {
            "type": "boolean",
            "title": "Scale",
            "default": false
          },
          "dim_reduce": {
            "type": "boolean",
            "title": "Dim Reduce",
            "default": false
          },
          "interval_finetune": {
            "type": "integer",
            "title": "Interval Finetune",
            "default": 5
          }
        },
        "type": "object",
        "title": "PreprocessingSchema"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}